refactored library not to use part/part of. fixed tests, benchmark and example to follow.
