+++ benchmark/benchmarks/mapping_benchmark.dart
@@ -1,3 +1,6 @@
+import 'package:benchmark_harness/benchmark_harness.dart';
+import 'package:robotlegs_di/src/injection/injector.dart';
+
 /*
 * Copyright (c) 2014 the original author or authors
 *
@@ -20,7 +23,6 @@
 * THE SOFTWARE.
 */
 
-part of robotlegs_di_benchmark;
 
 class MappingBenchmark extends BenchmarkBase {
   final int ITERATIONS = 1000;
@@ -31,17 +33,20 @@ class MappingBenchmark extends BenchmarkBase {
 
   void run() {
     for (int i = 0; i < ITERATIONS; i++) {
-      injector.getInstance(int, i.toString());
+      injector.getInstance(Abstract, i.toString());
     }
   }
 
   void setup() {
     for (int i = 0; i < ITERATIONS; i++) {
-      injector.map(int, i.toString()).toValue(i);
+      injector.map(Abstract, i.toString()).toValue(i);
     }
   }
 
   void teardown() {
     injector = null;
   }
+}
+
+abstract class Abstract {
 }
