--- lib/injectionpoints/constructor_injection_point.dart
@@ -1,48 +1,0 @@
-/*
-* Copyright (c) 2014 the original author or authors
-*
-* Permission is hereby granted, free of charge, to any person obtaining a copy
-* of this software and associated documentation files (the "Software"), to deal
-* in the Software without restriction, including without limitation the rights
-* to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-* copies of the Software, and to permit persons to whom the Software is
-* furnished to do so, subject to the following conditions:
-*
-* The above copyright notice and this permission notice shall be included in
-* all copies or substantial portions of the Software.
-*
-* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-* IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-* FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-* AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-* LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-* OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-* THE SOFTWARE.
-*/
-
-part of robotlegs_di;
-
-class ConstructorInjectionPoint extends MethodInjectionPoint {
-  //-----------------------------------
-  //
-  // Constructor
-  //
-  //-----------------------------------
-
-  ConstructorInjectionPoint(String method, List<dynamic> positionalArguments,
-      int numRequiredPositionalArguments, Map<String, dynamic> namedArguments)
-      : super(method, positionalArguments, numRequiredPositionalArguments,
-            namedArguments, false);
-
-  //-----------------------------------
-  //
-  // Public Methods
-  //
-  //-----------------------------------
-
-  dynamic createInstance(Type type, IInjector injector) {
-    ClassMirror reflectType = reflect.reflectType(type);
-    return reflectType.newInstance(
-        method, _getPositionalValues(injector, type, type));
-  }
-}
